Transcriber/requirements.txt
google-cloud-speech
google-cloud-storage
sounddevice
soundfile
tk
cryptography
